% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidiana.R
\name{Euclidiana}
\alias{Euclidiana}
\title{This function calculate the distance euclidiana}
\usage{
Euclidiana(Reals, Simulateds)
}
\arguments{
\item{Reals}{data frame with real data}

\item{Simulateds, }{data frame with simulated data}
}
\value{
Distance euclidiana between \code{Reals} and \code{Simulateds}.
}
\description{
This function calculate the distance euclidiana
}
