% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mahalanobis.R
\name{Mahalanobis}
\alias{Mahalanobis}
\title{This function calculate the distance Mahalabonis}
\usage{
Mahalanobis(Reals, Simulateds)
}
\arguments{
\item{Reals}{data frame with real data}

\item{Simulateds, }{data frame with simulated data}
}
\value{
Distance Mahalabonis between \code{Reals} and \code{Simulateds}.
}
\description{
This function calculate the distance Mahalabonis
}
