% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhattan.R
\name{Manhattan}
\alias{Manhattan}
\title{This function calculate the distance Manhattan}
\usage{
Manhattan(Reals, Simulateds)
}
\arguments{
\item{Reals}{data frame with real data}

\item{Simulateds, }{data frame with simulated data}
}
\value{
Distance Manhattan between \code{Reals} and \code{Simulateds}.
}
\description{
This function calculate the distance Manhattan
}
