% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariantes.R
\name{Univariantes}
\alias{Univariantes}
\title{Calcula Estadisticos Univariantes}
\usage{
Univariantes(Reales, Simulados, nombre_variable = NULL)
}
\arguments{
\item{Reales}{dataFrame datos Reles}

\item{Simulados}{dataFrame datos Simulados}

\item{nombre_variable}{vector de nombres de variables, opcional}
}
\value{
R^2, pearson, %RMSE, t-student, NS, MSEP, RMSEP, MAE, %MAE, U, MC, SC, RC, MV, AIC, bayes entre \code{Reales} and \code{Simulados}.
}
\description{
Calcula Estadisticos Univariantes
}
