% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariantes.R
\name{Univariantes}
\alias{Univariantes}
\title{This function calculate statistics Univariant}
\usage{
Univariantes(Reals, Simulateds, names_variables = NULL)
}
\arguments{
\item{Reals}{data frame with real data}

\item{Simulateds, }{data frame with simulated data}

\item{names_variables}{variable names vector, optional}
}
\value{
R^2, pearson, %RMSE, t-student, NS, MSEP, RMSEP, MAE, %MAE, U, MC, SC, RC, MV, AIC, bayes entre \code{Reals} and \code{Simulateds}.
}
\description{
This function calculate statistics Univariant
}
